% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_google_sheets.R
\name{read_google_sheets}
\alias{read_google_sheets}
\title{Read data from Google Sheets document}
\usage{
read_google_sheets(google_sheets_url, unnest_data = TRUE)
}
\arguments{
\item{google_sheets_url}{URL of the Google Sheets document.}

\item{unnest_data}{Logical, indicating whether to unnest the data. Default is TRUE.}
}
\value{
A tibble of tibbles, where each tibble corresponds to a sheet in the document.
The result includes sheet names and their corresponding data.
}
\description{
This function reads data from a Google Sheets document given its URL.
It uses the googlesheets4 package to fetch and parse the data from each sheet.
}
\examples{
\dontrun{
  google_sheets_url <- "https://docs.google.com/spreadsheets/d/1YsOKpxKOKlUEx6gAg0qO7JxJ52KWOZulmhwZYcGMYHY"
  read_google_sheets(google_sheets_url)
}

}
